"task62-доработка Country"

See merge request javapro_team29/social-net-team29!110

+++ common/src/main/java/ru/socialnet/team29/model/City.java
@@ -2,12 +2,11 @@ package ru.socialnet.team29.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 
 @Data
 @AllArgsConstructor
 public class City {
-    public Long id;
+    public Integer id;
     public String title;
 }
